Feature/23 (#24)
* feat: enhance package-level exports in service and utils

* refactor: simplify internal service imports and extract classes

* refactor: simplify imports in router and dto modules

* refactor: rename weighted_small_world_service to polynomial_optimization_service for clarity

* refactor: correct spelling of 'polynomial' in generator module and exports

* fix: rename flow_conservation_polynimial_generator to fix spelling typo

Agent-Logs-Url: https://github.com/quantum-guardians/mr2s-backend/sessions/dc6fb214-1f70-460c-bc10-4881c1b7a76e


* fix: rename flow_conservation_polynimial_generator to fix typo and update utils docstring

Agent-Logs-Url: https://github.com/quantum-guardians/mr2s-backend/sessions/f8d4f7a3-3244-475a-9473-96ae7562162e


* docs: update optimize docstring to describe PolynomialGenerator composition behavior

Agent-Logs-Url: https://github.com/quantum-guardians/mr2s-backend/sessions/28de2ce4-f7cb-4fb5-9d6f-5dfcdbfd6dce


* feat: add documentation for optimization system using polynomial generators

* Clarify abstract class description in documentation

Corrected the wording to clarify that the abstract class receives graph information.

---------
